+++ resume-analyser/next.config.ts
@@ -2,7 +2,8 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   /* config options here */
-  reactStrictMode: false
+  reactStrictMode: false,
+  devIndicators: false,
 };
 
 export default nextConfig;
